main_app/requirements.txt
fastapi
uvicorn
sqlalchemy
confluent_kafka
pydantic[email]
python-dotenv
